+++ node/packages/aws-lambda-otel-extension/internal/otel-extension-internal-node/wrapper.js
@@ -14,52 +14,60 @@ if (!EvalError.$serverlessHandlerFunction && !EvalError.$serverlessHandlerDeferr
 const awsLambdaInstrumentation = EvalError.$serverlessAwsLambdaInstrumentation;
 delete EvalError.$serverlessAwsLambdaInstrumentation;
 
-const wrapHandler = (handlerFunction) => {
+const wrapOriginalHandler = (originalHandler) => {
   let requestStartTime;
   let responseStartTime;
   let currentInvocationId = 0;
+  let isResolved = false;
+  const unresolvedPromise = new Promise(() => {});
   const debugLog = (...args) => {
     if (process.env.DEBUG_SLS_OTEL_LAYER) process._rawDebug(...args);
   };
+  let contextDone;
 
-  const wrappedHandler = awsLambdaInstrumentation._instance._getPatchHandler(
-    (event, context, callback) => {
+  const otelHandler = awsLambdaInstrumentation._instance._getPatchHandler(
+    (event, context, otelCallback) => {
       const invocationId = currentInvocationId;
       debugLog(
         'Extension overhead duration: internal request:',
         `${Math.round(Number(process.hrtime.bigint() - requestStartTime) / 1000000)}ms`
       );
 
-      const wrapCallback =
-        (originalCallback) =>
+      const wrapOtelCallback =
+        (someOtelCallback) =>
         (...args) => {
-          if (invocationId !== currentInvocationId) {
-            originalCallback(...args);
-            return;
-          }
+          // Callback invoked directly by Lambda logic, it'll invoke otel wrap callback
+          if (invocationId !== currentInvocationId) return;
+          if (isResolved) return;
+          isResolved = true;
           if (!responseStartTime) responseStartTime = process.hrtime.bigint();
-          originalCallback(...args);
+          someOtelCallback(...args);
         };
-      const done = wrapCallback(context.done);
-      context.done = done;
-      context.succeed = (result) => done(null, result);
-      context.fail = (err) => done(err == null ? 'handled' : err);
-      const result = handlerFunction(event, context, wrapCallback(callback));
+      contextDone = wrapOtelCallback(context.done);
+      context.done = contextDone;
+      context.succeed = (result) => contextDone(null, result);
+      context.fail = (err) => contextDone(err == null ? 'handled' : err);
+      const result = originalHandler(event, context, wrapOtelCallback(otelCallback));
       if (!result) return result;
       if (typeof result.then !== 'function') return result;
       return Promise.resolve(result).finally(() => {
-        if (invocationId !== currentInvocationId) return;
+        // Lambda logic resolved, passing result to otel wrapper
+        if (invocationId !== currentInvocationId) return unresolvedPromise;
+        if (isResolved) return unresolvedPromise;
+        isResolved = true;
         if (!responseStartTime) responseStartTime = process.hrtime.bigint();
+        return null;
       });
     }
   );
 
-  return (event, context, callback) => {
-    debugLog('Internal extension: Invocation');
-    const invocationId = ++currentInvocationId;
+  return (event, context, awsCallback) => {
     requestStartTime = process.hrtime.bigint();
+    EvalError.$serverlessInvocationStart = Date.now();
+    debugLog('Internal extension: Invocation');
+    isResolved = false;
+    ++currentInvocationId;
     const logResponseDuration = () => {
-      if (invocationId !== currentInvocationId) return;
       delete EvalError.$serverlessRequestHandlerPromise;
       delete EvalError.$serverlessResponseHandlerPromise;
       if (responseStartTime) {
@@ -70,35 +78,38 @@ const wrapHandler = (handlerFunction) => {
         responseStartTime = null;
       }
     };
-    const wrapCallback =
-      (originalCallback) =>
+    const wrapAwsCallback =
+      (someAwsCallback) =>
       (...args) => {
-        if (invocationId !== currentInvocationId) {
-          originalCallback(...args);
-          return;
-        }
+        // Callback invoked by Otel instrumentation after triggering response hook
         Promise.all([
           EvalError.$serverlessRequestHandlerPromise,
           EvalError.$serverlessResponseHandlerPromise,
         ]).finally(() => {
-          process.nextTick(() => originalCallback(...args));
+          process.nextTick(() => someAwsCallback(...args));
           logResponseDuration();
         });
       };
-    const done = wrapCallback(context.done);
-    context.done = done;
-    context.succeed = (result) => done(null, result);
-    context.fail = (err) => done(err == null ? 'handled' : err);
-    const result = wrappedHandler(event, context, wrapCallback(callback));
+    const originalDone = context.done;
+    contextDone = wrapAwsCallback(originalDone);
+    context.done = contextDone;
+    context.succeed = (result) => contextDone(null, result);
+    context.fail = (err) => contextDone(err == null ? 'handled' : err);
+    const result = otelHandler(event, context, wrapAwsCallback(awsCallback));
     if (!result) return result;
     if (typeof result.then !== 'function') return result;
-    return Promise.resolve(result).finally(() => {
-      if (invocationId !== currentInvocationId) return null;
-      return Promise.all([
-        EvalError.$serverlessRequestHandlerPromise,
-        EvalError.$serverlessResponseHandlerPromise,
-      ]).finally(logResponseDuration);
-    });
+    return Promise.resolve(result)
+      .finally(() => {
+        // Otel response hook triggered, passing result to AWS
+        return Promise.all([
+          EvalError.$serverlessRequestHandlerPromise,
+          EvalError.$serverlessResponseHandlerPromise,
+        ]).finally(logResponseDuration);
+      })
+      .finally(() => {
+        // AWS internally uses context methods to resolve promise result
+        contextDone = originalDone;
+      });
   };
 };
 
@@ -124,11 +135,11 @@ if (EvalError.$serverlessHandlerDeferred) {
     const handlerFunction = handlerContext[handlerFunctionName];
     if (typeof handlerFunction !== 'function') return handlerModule;
 
-    return { handler: wrapHandler(handlerFunction) };
+    return { handler: wrapOriginalHandler(handlerFunction) };
   });
   return;
 }
 
-const handlerFunction = EvalError.$serverlessHandlerFunction;
+const originalHandler = EvalError.$serverlessHandlerFunction;
 delete EvalError.$serverlessHandlerFunction;
-module.exports.handler = wrapHandler(handlerFunction);
+module.exports.handler = wrapOriginalHandler(originalHandler);