+++ node/packages/aws-lambda-otel-extension/test/fixtures/lambdas/callback-error.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports.handler = (event, context, callback) => {
+  callback(new Error('SLS Handled error'));
+};