+++ node/packages/aws-lambda-otel-extension/test/fixtures/lambdas/error-unhandled.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports.handler = () =>
+  setTimeout(() => {
+    throw new Error('SLS Unhandled error');
+  }, 500);